Moved PersonalStatementSummaryComponent
